Move lib app dirs and package include.
