+++ README.md
@@ -1,29 +1,3 @@
 # markserv-cli
+
 markserv command line interface
-
-- [test internal link](#test-ing)
-- [test markdown link](readme2.md)
-- [test markdown link 2](readme2)
-
-
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-<!--markserv|markdown|readme2.md-->
-
-# test-ing
-
-<!--markserv|markdown|readme2.md-->
-
