+++ index.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+module.exports = require('app/lib/markserv.js');
