+++ lib/core/argv.js
@@ -1,8 +1,8 @@
 const path = require('path');
 
-const pkg = require('app/../../package.json');
-const log = require('app/core/log');
-const helpFs = require('app/help/fs');
+const pkg = require('app/package.json');
+const log = require('app/lib/core/log');
+const helpFs = require('app/lib/help/fs');
 
 const cwd = process.cwd();
