+++ lib/core/service.js
@@ -1,15 +1,15 @@
 // const path = require('path');
 const Promise = require('bluebird');
 
-const log = require('app/core/log');
-const plugin = require('app/plugin/register');
-const help = require('app/help/plugins');
+const log = require('app/lib/core/log');
+const plugin = require('app/lib/plugin/register');
+const help = require('app/lib/help/plugins');
 
-const httpRequestHandler = require('app/http/requests');
-const httpServer = require('app/http/server');
-const templateCompiler = require('app/http/compiler');
+const httpRequestHandler = require('app/lib/http/requests');
+const httpServer = require('app/lib/http/server');
+const templateCompiler = require('app/lib/http/compiler');
 
-require('app/core/shutdown')(process);
+require('app/lib/core/shutdown')(process);
 
 const configureStack = (stack, Markconf) => {
   return new Promise((resolve, reject) => {
