+++ lib/core/shutdown.js
@@ -1,6 +1,6 @@
 const Sigint = require('sigint');
 
-const log = require('app/core/log');
+const log = require('app/lib/core/log');
 
 const sigint = Sigint.create();
