+++ lib/help/plugins.js
@@ -1,4 +1,4 @@
-const fs = require('app/help/fs');
+const fs = require('app/lib/help/fs');
 
 const configure = conf => {
   fs.configure(conf);
