+++ lib/http/compiler.js
@@ -3,8 +3,8 @@ const path = require('path');
 const Promise = require('bluebird');
 const cheerio = require('cheerio');
 
-const helpFs = require('app/help/fs');
-const log = require('app/core/log');
+const helpFs = require('app/lib/help/fs');
+const log = require('app/lib/core/log');
 
 let Markconf;
