+++ lib/http/requests.js
@@ -2,9 +2,9 @@ const path = require('path');
 
 const minimatch = require('minimatch');
 
-const log = require('app/core/log');
-const helpFs = require('app/help/fs');
-const templateCompiler = require('app/http/compiler');
+const log = require('app/lib/core/log');
+const helpFs = require('app/lib/help/fs');
+const templateCompiler = require('app/lib/http/compiler');
 
 let Markconf;
