+++ lib/http/server.js
@@ -6,7 +6,7 @@ const openport = require('openport');
 // const liveReload = require('livereload');
 // const connectLiveReload = require('connect-livereload');
 
-const log = require('app/core/log');
+const log = require('app/lib/core/log');
 
 let Markconf;
 let httpServer;
