+++ lib/markserv.js
@@ -1,13 +1,11 @@
-#!/usr/bin/env node
-
-const Markserv = require('app/core/service');
-const resolvePlugins = require('app/plugin/resolver');
+const Markserv = require('app/lib/core/service');
+const resolvePlugins = require('app/lib/plugin/resolver');
 
 const initialize = args => new Promise((resolve, reject) => {
-  const settings = require('app/core/argv').parse(args);
+  const settings = require('app/lib/core/argv').parse(args);
 
   resolvePlugins(settings.conf).then(plugins => {
-    const Markconf = require('app/core/markconf').initialize(settings, plugins);
+    const Markconf = require('app/lib/core/markconf').initialize(settings, plugins);
 
     Markserv.spawnService(Markconf)
     .then(resolve)
@@ -16,7 +14,7 @@ const initialize = args => new Promise((resolve, reject) => {
   .catch(reject);
 });
 
-const CLI = !module.parent;
+const CLI = !module.parent.parent;
 
 if (CLI) {
   initialize(process.argv);
