+++ lib/plugin/includer.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const fs = require('fs');
 
-const log = require('app/core/log');
+const log = require('app/lib/core/log');
 
 const filterModulePkgMeta = (pkg, moduleDir) => {
   const meta = {
