+++ lib/plugin/modifier.js
@@ -3,8 +3,8 @@ const fs = require('fs');
 
 const Promise = require('bluebird');
 
-const templateCompiler = require('app/http/compiler');
-const log = require('app/core/log');
+const templateCompiler = require('app/lib/http/compiler');
+const log = require('app/lib/core/log');
 
 const filterModulePkgMeta = (pkg, moduleDir) => {
   const meta = {
