+++ lib/plugin/register.js
@@ -1,10 +1,10 @@
 const fs = require('fs');
 const path = require('path');
 
-const registerModifier = require('app/plugin/modifier');
-const registerIncluder = require('app/plugin/includer');
-const help = require('app/help/plugins');
-const log = require('app/core/log');
+const registerModifier = require('app/lib/plugin/modifier');
+const registerIncluder = require('app/lib/plugin/includer');
+const help = require('app/lib/help/plugins');
+const log = require('app/lib/core/log');
 
 // File System
 module.exports.readfile = help.fs.readfile;
