+++ lib/plugin/resolver.js
@@ -2,9 +2,9 @@ const path = require('path');
 
 const Promise = require('bluebird');
 
-const log = require('app/core/log');
-const fs = require('app/help/fs');
-const plugin = require('app/plugin/register');
+const log = require('app/lib/core/log');
+const fs = require('app/lib/help/fs');
+const plugin = require('app/lib/plugin/register');
 // const verify = require('app/plugin/verifier');
 
 const loadModule = (confName, moduleName, dir, importDepth) => {
