+++ lib/plugin/verifier.js
@@ -1,4 +1,4 @@
-const log = require('app/core/log');
+const log = require('app/lib/core/log');
 
 const exception = (modifierName, should) => {
   const errorText = log.hl(modifierName) + ' ' + log.red(should);
