--- readme2.md
@@ -1,3 +1,0 @@
-## this is readme2.md
-
-<!--markserv|markdown|node_modules/markserv-contrib-app.github/partials/linked.md-->
